Extract helpful methods to the new tools module
